fix lr decay for training

+++ paddle/fluid/operators/distributed/request_handler_impl.cc
@@ -283,11 +283,11 @@ bool RequestNotifyHandler::Handle(const std::string& varname,
 
     auto* send_var = scope->FindVar(varname);
     auto send_var_tensor = send_var->Get<framework::LoDTensor>();
-    int64_t* send_value =
-        send_var_tensor.mutable_data<int64_t>(send_var_tensor.place());
+    float* send_value =
+        send_var_tensor.mutable_data<float>(send_var_tensor.place());
 
     auto counter = decay_counters.at(trainer_id);
-    counter += send_value[0];
+    counter += static_cast<int64_t>(send_value[0]);
     decay_counters.at(trainer_id) = counter;
 
     auto* global_step_var = this->scope()->FindVar(LEARNING_RATE_DECAY_COUNTER);
